More responsivenes and visual work
The projects and concepts sections now turn their items into a column on smaller screens.

Made the fade in assign all info to appear on a set timer if user is at the bottom of the page on refresh.

Otherwise the info appears as the person scrolls down the page a section at a time.