Hero banner layout and transitions
Messed around with the hero banner's sub title positioning since it seems to close to the second text.

Messed around with text color transitioning from one to another once a user hit a certain pixel number when scrolling. Contemplating wether I want this to actually be an automated thing or just a background jpg a user may scroll through and make the effect feel more dynamic.